refactor: organize by type
